+++ tests/xmlrs_reader_tests.rs
@@ -227,7 +227,7 @@ fn issue_83_duplicate_attributes() {
         r#"<hello><some-tag a='10' a="20"/></hello>"#,
         "
             |StartElement(hello)
-            |1:30 EmptyElement(some-tag, attr-error: error while parsing attribute: \
+            |1:30 EmptyElement(some-tag, attr-error: \
                   position 16: duplicated attribute, previous declaration at position 9)
             |EndElement(hello)
             |EndDocument